fix(FileBrowser): replace {local,remote}Directories with {local,remote}Hierarchy
Previously, {local,remote}Directories would hold all directories (recursively) of the
initial directory.  This is expensive to calculate for deep hierarchies. Also,
it was not really useful since it was calculated only once at the beginning.
This commit replaces it with {local,remote}Hierarchy a list of all parents of current
directory that can be use to go up , or in any parent through the drop-down like
in the Qt interface of a ParaView.